Fix some -Wshorten64-to-32 warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@39814 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/plus4/cart/magiccart.c
@@ -129,7 +129,7 @@ int magiccart_bin_attach(const char *filename, uint8_t *rawcart)
     if (fd == NULL) {
         return -1;
     }
-    len = util_file_length(fd);
+    len = (unsigned int)util_file_length(fd);
     fclose(fd);
 
     DBG(("magiccart_bin_attach len: %04x\n", len));
