+++ vice/src/plus4/cart/multicart.c
@@ -139,7 +139,7 @@ int multicart_bin_attach(const char *filename, uint8_t *rawcart)
     if (fd == NULL) {
         return -1;
     }
-    len = util_file_length(fd);
+    len = (unsigned int)util_file_length(fd);
     fclose(fd);
 
     DBG(("multicart_bin_attach len: %04x\n", len));
